Addictions can be configured in professions.json
