+++ data/raw/professions.json
@@ -5,7 +5,8 @@
   "description":
 "They say you don't have any 'useful skills', but at least your clothes fit\nwell.",
   "points": 0,
-  "items": ["jeans", "tshirt", "sneakers"] 
+  "items": ["jeans", "tshirt", "sneakers"],
+  "addictions": []
  },
  
  {
@@ -14,7 +15,8 @@
   "description":
 "You've made and adjusted clothing for some of the biggest names in town.\nNow, their undead bodies are trying to eat your brains. At least you'll be\nable to keep your own clothing repaired.",
   "points": 1,
-  "items": ["polo_shirt", "blazer", "pants", "dress_shoes", "sewing_kit", "scissors", "manual_tailor"]
+  "items": ["polo_shirt", "blazer", "pants", "dress_shoes", "sewing_kit", "scissors", "manual_tailor"],
+  "addictions": []
  },
 
  {
@@ -23,7 +25,8 @@
   "description":
 "It was your first day in the lab when the cataclysm hit. You were evacuated\nwith all the civilians, but at least you kept your lab gear.",
   "points": 1,
-  "items": ["dress_shirt", "pants", "boots", "coat_lab", "gloves_rubber", "glasses_safety"]
+  "items": ["dress_shirt", "pants", "boots", "coat_lab", "gloves_rubber", "glasses_safety"],
+  "addictions": []
  },
 
  {
@@ -32,7 +35,8 @@
   "description":
 "Although you never got your driver's license, you've always loved cars.\nYou start with a wrench, a toolbelt, and some car-related reading material.",
   "points": 2,
-  "items": ["tank_top", "jeans", "beltrig", "wrench", "mag_cars", "manual_mechanics"]
+  "items": ["tank_top", "jeans", "beltrig", "wrench", "mag_cars", "manual_mechanics"],
+  "addictions": []
  },
 
  {
@@ -41,6 +45,12 @@
   "description":
 "Everyone at work knew you as the person who always had a cigarette or two\nin hand. Now, you're down to a single pack, and you hope you find more soon.\nYou start out with a strong nicotine addiction, and 10 cigarettes.",
   "points": -1,
-  "items": ["pants", "dress_shirt", "dress_shoes", "cig"]
+  "items": ["pants", "dress_shirt", "dress_shoes", "cig"],
+  "addictions": [
+      {
+	  "type": "nicotine",
+	  "intensity": 10
+      }
+  ]
  }
 ]
